backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "riseagent-ai-backend"
version = "0.1.0"
description = "Production-ready multilingual voice agent backend for Rupeezy"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "chromadb>=0.5.0",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "langchain>=0.3.0",
  "langchain-anthropic>=0.3.0",
  "langchain-chroma>=0.2.0",
  "langchain-huggingface>=0.1.0",
  "langchain-mistralai>=0.2.0",
  "pydantic-settings>=2.6.0",
  "python-multipart>=0.0.12",
  "sentence-transformers>=3.0.0",
  "sqlalchemy>=2.0.0",
  "twilio>=9.3.0",
  "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
